requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic-settings
pytest
sqlalchemy
bcrypt
python-jose[cryptography]
email-validator
